refactor(kafka)!: fold kafka-connect-source into the kafka module
The Kafka Connect external source lived in its own module on the assumption that keeping it separate spared anyone pulling in plain Kafka from its extra dependencies. That assumption no longer holds: XT ships as three fat cloud images (aws/gcp/azure) that differ only in their provider SDKs, and those SDKs dwarf anything the Connect source adds. A consumer pulling in xtdb-kafka would not notice the connect-api, transforms, or converter jars against that backdrop, so the module boundary was buying isolation nobody was spending.

The source also already depended on xtdb-kafka (it resolves a !Kafka remote to a KafkaCluster), so the split was a higher layer folded over a lower one rather than two peers — collapsing it removes a layer rather than merging across one.

Package renamed xtdb.kafkaconnectsource -> xtdb.kafka.connectsrc to sit under the kafka module namespace, alongside the existing xtdb.kafka.proto. Error codes stay xtdb.kafka-connect-source/* — they are user-facing string identifiers, not bound to the Kotlin package, and stable codes avoid churn for anyone matching on them.

Converters are loaded reflectively from operator config (Class.forName on key.converter/value.converter); the source only ever names the abstract Converter interface. They are therefore never a compile dependency. connect-json was previously declared api despite this — corrected. Since image size is no longer the constraint, we bundle the two common converters (Json, Avro) as runtimeOnly so they work out of the box; operators wanting JsonSchema or others add the jar to their own classpath. The Confluent serializers and the json-schema converter stay test-only — tests reference them as config strings to produce messages, not as bundled runtime behaviour.

Behaviour is preserved: the moved files differ from their originals only by the package token, and the three docker images drop the now-redundant xtdb-kafka-connect-source dependency since the code lives in xtdb-kafka they already pull in. The artifact rename and proto type-URL change would be breaking had this shipped, but the source is unreleased — no published artifact, no persisted external-source config — so nothing downstream breaks.
