--- modules/kafka-connect-source/src/main/resources/META-INF/services/xtdb.database.ExternalSource$Registration
@@ -1,1 +1,0 @@
-xtdb.kafkaconnectsource.KafkaConnectSource$Factory$Registration
