--- modules/kafka-connect-source/src/main/resources/META-INF/services/xtdb.kafkaconnectsource.RecordIndexer$Registration
@@ -1,1 +1,0 @@
-xtdb.kafkaconnectsource.DocsIndexer$Factory$Registration
