+++ modules/kafka/src/main/kotlin/xtdb/kafka/connectsrc/RecordIndexer.kt
@@ -1,4 +1,4 @@
-package xtdb.kafkaconnectsource
+package xtdb.kafka.connectsrc
 
 import kotlinx.serialization.modules.PolymorphicModuleBuilder
 import kotlinx.serialization.modules.SerializersModule
