+++ modules/kafka/src/main/resources/META-INF/services/xtdb.database.ExternalSource$Registration
@@ -1,0 +1,1 @@
+xtdb.kafka.connectsrc.KafkaConnectSource$Factory$Registration
