+++ modules/kafka/src/main/resources/META-INF/services/xtdb.kafka.connectsrc.RecordIndexer$Registration
@@ -1,0 +1,1 @@
+xtdb.kafka.connectsrc.DocsIndexer$Factory$Registration
